feat: moved test helpers into shared namespace
